refactor: split tests.rs (2,851 lines) into 42 focused test modules
Split monolithic tests.rs into tests/ directory with modules organized
by feature area. Each module is under 100 lines.

Categories:
- CRUD: crud.rs
- Filtering: where_clause.rs, like.rs, between.rs, distinct.rs
- DML: update.rs, delete.rs, insert_select.rs
- DDL: ddl.rs, ddl_ext.rs
- Expressions: expressions.rs, cast.rs, coalesce.rs, case_when.rs
- Functions: string_func.rs, math_func.rs
- Aggregates: aggregate.rs, aggregate_ext.rs
- JOINs: join.rs, join_regression.rs, join_null.rs
- Subqueries: subquery.rs, subquery_ext.rs
- Set ops: set_ops.rs
- ORDER/LIMIT: order_limit.rs
- Constraints: constraints.rs
- RETURNING: returning.rs
- SERIAL: serial.rs
- Vectors: vector.rs, vector_hnsw.rs, vector_hnsw_recall.rs
- Windows: window.rs, window_value.rs, window_aggregate.rs,
  window_running.rs, window_partition.rs
- CTEs: cte.rs, cte_ext.rs
- UPSERT: upsert.rs, upsert_ext.rs
- Cross-table: cross_table.rs
- Misc: misc.rs

All 236 tests pass. No test logic modified.
